refactor(cli): Improve `pr` module organization
